Fix a lot of things to at least get it to synthesize
Most significant change is probably that we now just MUX together the
various AXI4 channels to avoid having to use vivado IP cores (which are
more performant indeed but we don't need maximum performance and are a
pain to handle with the 5000 line wrappers).
No idea if any of this works.